requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.8
autogen-agentchat==0.5.3
autogen-core==0.5.3
autogen-ext==0.5.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Deprecated==1.2.18
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
jiter==0.9.0
jsonref==1.1.0
openai==1.75.0
opentelemetry-api==1.32.1
packaging==25.0
pillow==11.2.1
protobuf==5.29.4
pyautogen==0.8.7
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
pywin32==310
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
termcolor==3.0.1
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
wrapt==1.17.2
zipp==3.21.0
